+++ test/scripts/test_lav.py
@@ -36,11 +36,19 @@ class TestLAVScript(unittest.TestCase):
             check_call(['tfaip-lav',
                         '--export_dir', os.path.join(d, 'best'),
                         '--pipeline.limit', '10',
+                        ])
+            check_call(['tfaip-lav',
+                        '--export_dir', os.path.join(d, 'best'),
+                        '--pipeline.limit', '10',
+                        '--data',
+                        'examples.tutorial.full.data.prediction_data_generation:TutorialPredictionGeneratorParams',
                         '--data.files', os.path.join(work_dir, 'data', '*.png'),
                         ])
             check_call(['tfaip-lav',
                         '--export_dir', os.path.join(d, 'best'),
                         '--pipeline.limit', '10',
+                        '--data',
+                        'examples.tutorial.full.data.prediction_data_generation:TutorialPredictionGeneratorParams',
                         '--data.files', os.path.join(work_dir, 'data', '*.png'),
                         '--run_eagerly',
                         '--dump', os.path.join(d, 'dump.pkl'),
