### 2021-04-27 #3142 @arfon
@ankur-gupta & @Het-Shah - would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission we'd be asking you to review is [_tfaip - a Generic and Powerful Research Framework for Deep Learning based on Tensorflow_](https://github.com/Planet-AI-GmbH/tfaip).
