+++ tfaip/scenario/listfile/listfilelav.py
@@ -25,7 +25,7 @@ from tfaip.scenario.listfile.params import ListsFileGeneratorParams
 class ListFileLAV(LAV):
     def _on_lav_end(self, data_generator_params: ListsFileGeneratorParams, result):
         if self._params.store_results:
-            dump_dict = {"metrics": {k: v for k, v in result.items() if not isinstance(v, bytes)},
+            dump_dict = {"metrics": {k: v for k, v in result.items() if (not isinstance(v, bytes) and not type(v).__module__ == 'numpy')},
                          "lav_params": self._params.to_dict(),
                          "data_params": self._data.params.to_dict(),
                          "model_params": self._model.params.to_dict()}
