+++ tfaip/trainer/callbacks/benchmark_callback.py
@@ -16,91 +16,151 @@
 # tfaip. If not, see http://www.gnu.org/licenses/.
 # ==============================================================================
 """Definition of the BenchmarkCallback"""
-from tensorflow.keras.callbacks import Callback
-from time import time, strftime, gmtime
 import logging
+from copy import copy
+from dataclasses import dataclass
+
+import prettytable
+from tensorflow.keras.callbacks import Callback
+
+from tfaip.util.profiling import MeasureTime
 
 logger = logging.getLogger(__name__)
 
 
-class RunningAverage:
-    """Compute the running average value"""
-    def __init__(self):
-        self.steps = 0
-        self.avg = 0
+@dataclass
+class BenchmarkResults:
+    """Class storing benchmark results of training or testing or prediction
 
-    def add(self, v):
-        self.avg = (self.avg * self.steps + v) / (self.steps + 1)
-        self.steps += 1
+    Use pretty_print to print a formatted table of the full results.
+    """
+    n_batches: float = 0
+    n_samples: float = 0
+    total_time: float = 1e-10
+    avg_time_per_batch: float = 0
+    avg_time_per_sample: float = 0
+    batches_per_second: float = 0
+    samples_per_second: float = 0
 
-    def reset(self):
-        self.steps = 0
-        self.avg = 0
+    def pretty_print(self, print_fn=print):
+        table = prettytable.PrettyTable(['', 'Total', 'Batch', 'Sample'])
+        table.add_row(['Count', 1, self.n_batches, self.n_samples])
+        table.add_row(['Time Per', self.total_time, self.avg_time_per_batch, self.avg_time_per_sample])
+        table.add_row(['Per Second', 1 / self.total_time, self.batches_per_second, self.samples_per_second])
+        if print_fn is None:
+            return table
+        print_fn(table)
+
+    def finish_epoch(self, duration, samples=None):
+        self.total_time = duration
+        if samples is not None:
+            self.n_samples = samples
+        if self.n_batches > 0:
+            self.avg_time_per_batch /= self.n_batches
+            self.avg_time_per_sample /= self.n_samples
+            self.batches_per_second = 1 / self.avg_time_per_batch
+            self.samples_per_second = 1 / self.avg_time_per_sample
+        return self
+
+    def finish_batch(self, batch_size, batch_time):
+        self.n_batches += 1
+        if batch_size is not None:
+            self.n_samples += batch_size
+        self.avg_time_per_batch += batch_time
+        self.avg_time_per_sample += batch_time
+        return self
 
 
 class BenchmarkCallback(Callback):
     """
     The BenchmarkCallback will trace the training and validation times per patch, epoch, and in total.
     """
+
     def __init__(self):
         super().__init__()
-        self._supports_tf_logs = True       # Any Callback before LAV callback must act on raw tf logs only
-        self.train_start_time = -1
-        self.epoch_start_time = -1
-        self.batch_start_time = -1
-        self.test_start_time = -1
+        self._supports_tf_logs = True  # Any Callback before LAV callback must act on raw tf logs only
 
-        self.avg_time_per_batch = RunningAverage()
-        self.avg_time_per_epoch = RunningAverage()
-        self.avg_time_per_test = RunningAverage()
+        self.total_train_time = MeasureTime()
+        self.total_epoch_time = MeasureTime()
+        self.total_test_time = MeasureTime()
+        self.batch_time = MeasureTime()
 
-        self.last_dt_per_epoch = -1
-        self.last_dt_per_test = -1
-        self.last_dt_per_batch = RunningAverage()
[…]
