### 2021-05-20 #3297 @whedon
[REVIEW]: tfaip - a Generic and Powerful Research Framework for Deep Learning based on Tensorflow
**Submitting author:** @chwick (<a href="http://orcid.org/0000-0003-3958-6240">Christoph Wick</a>)
**Repository:** <a href="https://github.com/Planet-AI-GmbH/tfaip" target ="_blank">https://github.com/Planet-AI-GmbH/tfaip</a>
**Version:** v1.1.0
