### 2021-05-20 #3297 @levimcclenny
I'm trying to accept the invite to edit the checklist but its saying "invite expired" - Any idea?

Thanks!
