### 2021-05-21 Planet-AI-GmbH/tfaip#3 @Het-Shah
Review for your JOSS submission
Hi @ChWick! I am a reviewer for your work at JOSS ([link](https://github.com/openjournals/joss-reviews/issues/3297)). Overall, the structure of the library seems great to me!  However, I had few concerns which I would like to add here 😄. I believe that these won't take a lot of time to rectify.

Concerns regarding the repository -
- [ ] Can you benchmark on any one scenario? For example, min/full tutorial can be benchmarked on a simple model architecture from TensorFlow.
- [ ] Can you add more tutorials to this repository? I can find an additional repository that has different scenarios, [link](https://github.com/Planet-AI-GmbH/tfaip_example_scenarios). Having it in one repository makes it easier for end-user to use them for their work directly
- [ ] Can you add the link to Contributor guidelines (mentioned in the [documentation](https://tfaip.readthedocs.io/en/latest/doc.development.html#guidelines)) in the main README.md file. [Sorry for asking you to do this 😅]

Concerns regarding the paper -
- [ ] In the section "State of the Field," compare _tfaip_ with some other packages based on TensorFlow and provide advantages of tfaip over them.
- [ ] Remove all the hyperlinks and add them in either the reference section or in footnotes.

You can let me know if you have any arguments against any of the concerns I have raised. I am more than happy to discuss these. 😄
I'll add more concerns as and when I find them!
