### 2021-05-24 Planet-AI-GmbH/tfaip#3 @levimcclenny
Hey @ChWick, I'm the second reviewer for JOSS.

In addition to the above, I believe a short example in the text would be useful. This paper does something similar - https://joss.theoj.org/papers/10.21105/joss.03077

Also for comparison, I would list how this compares to things like AutoKeras, etc. so the users can see how this package fits into the AutoML pipeline. That would be useful for me (for example) as someone familiar, but not well-versed in the topic. I am interested in using AutoML in projects of my own and I believe a comparison or discussion about this packages capabilities vs. other state-of-the-art packages would be welcome so I can see how I may do so.

Other than that the paper feels very abbreviated, but comparing to other JOSS papers I believe this is somewhat standard. Most others don't use bullets (from looking at a few of the recent ones on the page), so reconfiguring into a paragraph format could be good to make it more as a "journal paper" or application note style.

The software itself is pretty good, and very timely. I think the documentation is very professional. I agree that providing more about contribution guidelines would be welcome. You have how people can build a local version and change stuff, but no guidelines on how a user would submit a PR, for instance, or what you are interested in developing in the future for others to grab on to.
