### 2021-05-24 #3297 @levimcclenny
I also commented on the issue opened by @Het-Shah. I think it's a good paper as well. A few minor things but overall well-written and the documentation is really professional. Would like to see more accessible examples though. Other comments are on the repo.
