### 2021-05-25 Planet-AI-GmbH/tfaip#3 @ChWick
@Het-Shah @levimcclenny Thank you so much for reviewing the paper, and - of course - the software! We really appreciate your efforts and time investments!
Also thanks for the two reasonable initial reviews! We will adopt the paper and repository accordingly. This will take some time, I will reply after everything is done!

Some specific questions @Het-Shah:
* What do you mean by "benchmark". Do you propose a "speed" or "accuracy" comparison to Tensorflow's MNIST-Tutorial, or both?
* Do you think that merging the "examples" and "base" tfaip repositories suffices to provide additional "tutorials", or do you recommend additional ones?
