### 2021-05-25 Planet-AI-GmbH/tfaip#3 @Het-Shah
> What do you mean by "benchmark". Do you propose a "speed" or "accuracy" comparison to Tensorflow's MNIST-Tutorial, or both?

I would personally prefer both and a mention of the hardware used (can be a GPU from colab as well). You can easily collect these metrics in one run, so shouldn't be a problem in my opinion. If you are already in process of running experiments just a comparison of validation accuracies will also suffice.

>Do you think that merging the "examples" and "base" tfaip repositories suffices to provide additional "tutorials", or do you recommend additional ones?

I think merging those repositories is fine for this submission, but if in the future you want to add more examples it's totally up to you.
