### 2021-05-31 188434b
Including suggestions of reviewers for joss paper (#3):
* Merged with tfaipscenarios which are now part of the examples
* Added contribution docs
* Added MNIST and data pipeline benchmarks
* Updated paper
