+++ joss/paper.md
@@ -43,24 +43,26 @@ Each scenario is highly configurable by parameters that can directly be modified
 
 # Statement of Need
 
-An application of _tfaip_ resolves recurrent obstacles of research and development in an elegant and robust way:
+An application of _tfaip_ resolves recurrent obstacles of research and development in an elegant and robust way.
 
-* A complete scenario including the graph (e.g., the network architecture), the training (e.g., the optimizer or learning rate schedule), and the data pipeline (e.g., the data sources) are fully parameterizable by the command line or the API.
-  This simplifies hyper-parameter-optimization but also testing of novel ideas.
-* Each scenario is created by implementing predefined interfaces (e.g., loss-function or the graph construction).
-  This leads to a clean, structured, modularized, and readable code preventing bad code style and facilitates maintenance.
-* _tfaip_ provides a simple API to deploy a scenario.
-  The corresponding module will automatically apply pre-processing, infer the trained model, and optionally transform the output by a post-processing pipeline.
-  The information about the pipeline-construction is embedded within the model which enables to store and load models with a different data pipeline even for the same scenario.
-  This is handy if, for example, certain pre-processing steps are not required for one specific model or other inputs are expected.
-* During research, a tedious step is data preparation which often comprises the conversion of data into the format required by the framework.
-  Tensorflow allows integrating Python code in the data pipeline which is however not run (truly) in parallel by multiple processes and results quite often in a bottleneck.
-  To speed-up Tensorflow, a user has to transform Python into Tensorflow operations which is laborious, partly even impossible, and complicates debugging.
-  _tfaip_ tackles this issue by providing a sophisticated pipeline setup based on so-called data processors which apply simple transformation operations in pure Python code and are automatically executed in parallel.
+A complete scenario including the graph (e.g., the network architecture), the training (e.g., the optimizer or learning rate schedule), and the data pipeline (e.g., the data sources) are fully parameterizable by the command line or the API.
+This simplifies hyper-parameter-optimization but also testing of novel ideas.
+Each scenario is created by implementing predefined interfaces (e.g., loss-function or the graph construction).
+This leads to a clean, structured, modularized, and readable code preventing bad code style and facilitates maintenance.
+
+_tfaip_ provides a simple API to deploy a scenario.
+The corresponding module will automatically apply pre-processing, infer the trained model, and optionally transform the output by a post-processing pipeline.
+The information about the pipeline-construction is embedded within the model which enables to store and load models with a different data pipeline even for the same scenario.
+This is handy if, for example, certain pre-processing steps are not required for one specific model or other inputs are expected.
+
+During research, a tedious step is data preparation which often comprises the conversion of data into the format required by the framework.
+Tensorflow allows integrating Python code in the data pipeline which is however not run (truly) in parallel by multiple processes and results quite often in a bottleneck.
+To speed-up Tensorflow, a user has to transform Python into Tensorflow operations which is laborious, partly even impossible, and complicates debugging.
+_tfaip_ tackles this issue by providing a sophisticated pipeline setup based on so-called data processors which apply simple transformation operations in pure Python code and are automatically executed in parallel.
 
 # State of the Field
 
-Efficient research in the area of Deep Learning requires the integration of highly sophisticated Open-Source frameworks such as [Tensorflow](https://www.tensorflow.org/) [@tensorflow2015-whitepaper], [PyTorch](https://pytorch.org/) [@pytorch_2019], [Caffe](https://github.com/BVLC/caffe) [@jia2014caffe], [CNTK](https://github.com/microsoft/CNTK) [@cntk2016], or [Trax](https://github.com/google/trax) [@trax2021].
[…]
