+++ joss/paper.bib
@@ -9,7 +9,7 @@
 
 @inproceedings{wick_bidirectional_2021,
 	title = {{Bidirectional Transformer for Handwritten Text Recognition}},
-	volume = {submitted to},
+	volume = {accepted for},
 	booktitle = {16th {International} {Conference} on {Document} {Analysis} and {Recognition} ({ICDAR})},
 	publisher = {IEEE},
 	author = {Wick, Christoph and Zöllner, Jochen and Grüning, Tobias},
@@ -97,4 +97,19 @@ doi = {10.1145/2939672.2945397}
   publisher = {GitHub},
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/google/trax}},
+}
+
+@misc{pycharm2021,
+  author = {{JetBrains}},
+  title = {PyCharm},
+  year = {2021},
+  howpublished = {\url{https://jetbrains.com/pycharm}}
+}
+
+@article{zoellner2021,
+	title = {Optimizing small BERTs trained for German NER},
+	journal = {Computational Linguistics},
+	volume={submitted to},
+	author = {Zöllner, Jochen and Sperfeld, Konrad and Wick, Christoph and Labahn, Roger},
+	year = {2021}
 }
