### 2021-05-31 Planet-AI-GmbH/tfaip#3 @ChWick
Hi, we included you suggestions:

@Het-Shah

* We added a brief benchmark in the README.md for MNIST and the data pipeline.
*  We merged the examples which is certainly beneficial!
*  We added [contribution guidelines](https://tfaip.readthedocs.io/en/latest/doc.development.html) to the docs.
*  We added a comparison to AutoML (do you have other suggestions for packages to compare to?)
*  We removed all optional hyperlinks.

@levimcclenny
We included a paragraph about AutoML and reformatted the paper to be more "paper-like" (no bullet lists). You suggested to include a short example. Do you mean to include the [tutorial code](https://github.com/Planet-AI-GmbH/tfaip/tree/master/examples/tutorial/min) which is probably too long, or the "training calls" as seen in the [README.md](https://github.com/Planet-AI-GmbH/tfaip/blob/master/README.md)?
