### 2021-06-04 #3297 @arfon
@ChWick – I've found a small issue with your paper here that makes it a little more readable: https://github.com/Planet-AI-GmbH/tfaip/pull/4

Also, I find the current 'Statement of Need' section rather confusing to read. This section should clearly state for the reader why this software is needed, what problems it uniquely solves (or solves better than other tools) etc. Some comments on this section:

This sentence seems out of context?

> An application of tfaip resolves recurrent obstacles of research and development in an elegant and robust way.

Are you saying that other tools don't enable this?

> A complete scenario including the graph (e.g., the network architecture), the training (e.g., the optimizer or learning rate schedule),...

This next paragraph again is lacking clarity. Is _tfaip_ unique at being able to describe the 'complete scenario'? What is hard about how current tools do this and how is _tfaip_ solving these problems?

> tfaip provides a simple API to deploy a scenario. The corresponding module will automatically apply pre-processing, infer the trained model, and optionally transform the output by a post-processing pipeline.

Again, for the next paragraph, what is the value-add here compared to other tools or challenges data practitioners currently face?
