### 2021-06-18 #3297 @ChWick
@arfon Thank you very much for this comment! We agree that this section required some modification which I guess is also in the interest of the other two reviewers!
We completely rewrote the section in order to clarify but also highlight the "Need" of _tfaip_. Please let us know if you have additional remarks.

Please also see Planet-AI-GmbH/tfaip#4.
