### 2021-06-21 f5cb7d8
A few extra improvements to readability. (#6)
* Update paper.md

* Update joss/paper.md
