+++ tfaip/version.py
@@ -15,4 +15,4 @@
 # You should have received a copy of the GNU General Public License along with
 # tfaip. If not, see http://www.gnu.org/licenses/.
 # ==============================================================================
-__version__ = "1.2.1"
+__version__ = "1.2.2"
